Move non-manager modules into a utils package.
Utils is a bad name, but I can't think of anything better right now.
