+++ qaz_modules/brew.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
 
-from qaz.managers import shell
 from qaz.modules.base import Module
 from qaz.modules.registry import registry
+from qaz.utils import shell
 
 
 @registry.register
