+++ qaz_modules/iterm2.py
@@ -2,9 +2,9 @@ from __future__ import annotations
 
 import sys
 
-from qaz.managers import shell
 from qaz.modules.base import Module
 from qaz.modules.registry import registry
+from qaz.utils import shell
 
 
 class ITerm2(Module):
