+++ qaz_modules/vim.py
@@ -2,10 +2,10 @@ from __future__ import annotations
 
 from pathlib import Path
 
-from qaz.managers import git
-from qaz.managers import shell
 from qaz.modules.base import Module
 from qaz.modules.registry import registry
+from qaz.utils import git
+from qaz.utils import shell
 
 
 @registry.register
