+++ qaz_modules/zsh.py
@@ -4,10 +4,10 @@ import sys
 from pathlib import Path
 
 from qaz.managers import brew
-from qaz.managers import git
-from qaz.managers import shell
 from qaz.modules.base import Module
 from qaz.modules.registry import registry
+from qaz.utils import git
+from qaz.utils import shell
 
 
 class MacOSZsh(Module):
